naming changes, commenting
